Example files and working PONE- ADD-perturbations
Example files library extended

Insert one perturbation point for an add instruction according to the PONE
scheme. The point is when the  binary is run activated at random. I.E a
binary with one single pp is generated. the pp wil be activated at random.
THe probability of activiating the pp is set in pone.c