Conduct attacks on Chess2016 and kryptologik
In this folder will all the attack related scrips be located.

As of now it is possible to conduct the regular side channel provided
in the deadpool repo on chess2016 and kryptologik. NSC is up next.
A statistics module to parse the loggs and look at them in more depth
is provided. A scoring system based on the number of correctly guesed
bytes is implemented.